Restructured project and splitted code to modules to enable React Hot Reloading
